backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary
whitenoise
dj-database-url
gspread==6.2.1
google-auth==2.49.1
google-auth-oauthlib==1.3.0
openpyxl==3.1.5
requests==2.33.0
PyJWT==2.12.1
sqlparse==0.5.5
urllib3==2.6.3
gunicorn
